requirements-test.txt
coverage==3.2b3
nose==1.0.0
nosexcover==1.0.4
django-nose==0.1.2
pylint==0.23.0
lxml==2.3
lettuce==0.1.20
